create common test pkg. convert TestGet, TestList
