+++ test/doc.go
@@ -1,0 +1,2 @@
+// Package test contains tests for DBs. Implementation under test is pluggable.
+package test
